Integrate LightboxViewController in Guteberg
