+++ WordPress/Classes/Utility/Media/ImageLoader.swift
@@ -83,18 +83,13 @@ import WordPressMedia
     @objc(loadImageWithURL:fromPost:preferredSize:placeholder:success:error:)
     func loadImage(with url: URL, from post: AbstractPost, preferredSize size: CGSize = .zero, placeholder: UIImage?, success: ImageLoaderSuccessBlock?, error: ImageLoaderFailureBlock?) {
 
-        let host = MediaHost(with: post, failure: { error in
-            WordPressAppDelegate.crashLogging?.logError(error)
-        })
-
+        let host = MediaHost(post)
         loadImage(with: url, from: host, preferredSize: size, placeholder: placeholder, success: success, error: error)
     }
 
     @objc(loadImageWithURL:fromReaderPost:preferredSize:placeholder:success:error:)
     func loadImage(with url: URL, from readerPost: ReaderPost, preferredSize size: CGSize = .zero, placeholder: UIImage?, success: ImageLoaderSuccessBlock?, error: ImageLoaderFailureBlock?) {
-
-        let host = MediaHost(with: readerPost)
-        loadImage(with: url, from: host, preferredSize: size, placeholder: placeholder, success: success, error: error)
+        loadImage(with: url, from: MediaHost(readerPost), preferredSize: size, placeholder: placeholder, success: success, error: error)
     }
 
     /// Load an image from a specific post, using the given URL. Supports animated images (gifs) as well.
