+++ WordPress/Classes/ViewRelated/Blaze/Overlay/BlazePostPreviewView.swift
@@ -96,13 +96,9 @@ final class BlazePostPreviewView: UIView {
 
         if let url = post.featuredImageURL {
             featuredImageView.isHidden = false
-            let host = MediaHost(with: post, failure: { error in
-                // We'll log the error, so we know it's there, but we won't halt execution.
-                WordPressAppDelegate.crashLogging?.logError(error)
-            })
             let preferredSize = CGSize(width: featuredImageView.frame.width, height: featuredImageView.frame.height)
                 .scaled(by: UITraitCollection.current.displayScale)
-            featuredImageView.setImage(with: url, host: host, size: preferredSize)
+            featuredImageView.setImage(with: url, host: MediaHost(post), size: preferredSize)
 
         } else {
             featuredImageView.isHidden = true
