+++ WordPress/Classes/ViewRelated/Post/Views/PostCompactCell.swift
@@ -78,11 +78,7 @@ final class PostCompactCell: UITableViewCell, Reusable {
         if let post, let url = post.featuredImageURL {
             featuredImageView.isHidden = false
 
-            let host = MediaHost(with: post, failure: { error in
-                // We'll log the error, so we know it's there, but we won't halt execution.
-                WordPressAppDelegate.crashLogging?.logError(error)
-            })
-
+            let host = MediaHost(post)
             let targetSize = Constants.imageSize.scaled(by: traitCollection.displayScale)
             featuredImageView.setImage(with: url, host: host, size: targetSize)
         } else {
