+++ WordPress/Classes/ViewRelated/Post/Views/PostListCell.swift
@@ -75,9 +75,7 @@ final class PostListCell: UITableViewCell, AbstractPostListCell, PostSearchResul
         featuredImageView.isHidden = viewModel.imageURL == nil
         featuredImageView.layer.opacity = viewModel.syncStateViewModel.isEditable ? 1 : 0.25
         if let imageURL = viewModel.imageURL {
-            let host = MediaHost(with: viewModel.post) { error in
-                WordPressAppDelegate.crashLogging?.logError(error)
-            }
+            let host = MediaHost(viewModel.post)
             let thumbnailURL = MediaImageService.getResizedImageURL(for: imageURL, blog: viewModel.post.blog, size: Constants.imageSize.scaled(by: UIScreen.main.scale))
             featuredImageView.setImage(with: thumbnailURL, host: host)
         }
